add product details bottom sheet

+++ lib/feature/newProduct/model/response/product_response.dart
@@ -46,7 +46,7 @@ class PaginationRuslt {
 class DataProductResponse {
   String? title;
   String? description;
-  int? ratingsAverage;
+  double? ratingsAverage;
   @JsonKey(name: "_id")
   String? sId;
   int? price;
